Move TrackPiece control to `Controls` folder.
