+++ TrackDesigner/Controls/TrackPiece.xaml
@@ -1,16 +1,16 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:trackDesigner="clr-namespace:TrackDesigner">
+                    xmlns:controls="clr-namespace:TrackDesigner.Controls">
 <ResourceDictionary.MergedDictionaries>
     <ResourceDictionary>
         <RotateTransform x:Key="FlipOverTransform" Angle="180" CenterX="49" CenterY="49"/>
     </ResourceDictionary>
 </ResourceDictionary.MergedDictionaries>
 
-    <Style x:Key="TrackPieceStyle" TargetType="{x:Type trackDesigner:TrackPiece}">
+    <Style x:Key="TrackPieceStyle" TargetType="{x:Type controls:TrackPiece}">
         <Setter Property="Template">
             <Setter.Value>
-                <ControlTemplate TargetType="{x:Type trackDesigner:TrackPiece}">
+                <ControlTemplate TargetType="{x:Type controls:TrackPiece}">
                     <Border x:Name="Part_Border" 
                             Background="White" 
                             BorderBrush="DarkGray"
