+++ TrackDesigner/MainWindow.xaml.cs
@@ -2,6 +2,7 @@
 using System.Windows.Controls;
 using System.Windows.Controls.Ribbon;
 using System.Windows.Media;
+using TrackDesigner.Controls;
 
 namespace TrackDesigner
 {
