+++ TrackDesigner/Themes/Generic.xaml
@@ -1,10 +1,10 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:trackDesigner="clr-namespace:TrackDesigner">
+                    xmlns:controls="clr-namespace:TrackDesigner.Controls">
 
     <ResourceDictionary.MergedDictionaries>
-        <ResourceDictionary Source="pack://application:,,,/TrackDesigner;component/Themes/TrackPiece.xaml" />
+        <ResourceDictionary Source="pack://application:,,,/TrackDesigner;component/Controls/TrackPiece.xaml" />
     </ResourceDictionary.MergedDictionaries>
 
-    <Style BasedOn="{StaticResource TrackPieceStyle}" TargetType="trackDesigner:TrackPiece"/>
+    <Style BasedOn="{StaticResource TrackPieceStyle}" TargetType="controls:TrackPiece"/>
 </ResourceDictionary>
